An "expand" property has been added to EasyLayoutRow to control how items are filled in a row.
**Breaking change:** EasyLayoutFlexible.flex is now int instead of double.

+++ package/lib/src/easy_layout_flexible.dart
@@ -8,7 +8,7 @@ import 'package:flutter/widgets.dart';
 ///
 class EasyLayoutFlexible extends StatelessWidget {
   final Widget child;
-  final double flex;
+  final int flex;
 
   const EasyLayoutFlexible({
     Key key,
